Fix crash, data-corruption, and correctness bugs from July 2026 code review
Full-codebase review fixes (Phases 1-4 of the improvement plan). All
changes affect both mono and color builds unless noted.

Crash / data corruption:
- Guard active_session in the keyDown path: menu dispatch (Cmd+Q, Close
  on last window) can destroy sessions mid-keystroke; the loop then
  dereferenced NULL, writing low memory before an address error
- Rewrite prefs migration with frozen per-era on-disk layouts in new
  settings_migrate.c: the old in-place ladder aliased historical bytes
  onto current struct offsets, corrupting bookmarks/fonts/DNS/username
  for every pre-v14 prefs file. v8-era DNS server is now recovered.
  Covered by native test suite (tests/migration_test.c, 112 checks)
- Clamp log header/footer lengths: long hostnames leaked stack bytes
  into session logs (logging builds)

Correctness:
- ESC H (HTS) left parser stuck in escape state, swallowing next byte
- Insert mode (CSI 4h) was bypassed by the ASCII fast path
- Saturated truecolor mapped to gray: 16-bit overflow in 256-color
  distance math (color builds, System 7)
- DNS-over-TCP fallback now loops until the full message arrives
  instead of failing on multi-segment responses
- Paste converts CR to NVT-standard CR LF (clipboard builds)
- Remove Cmd+digit F-key aliases so Cmd+1-0 switch sessions as the
  Window menu advertises; F1-F12 still work via function-key vkeys
- Find in Scrollback matches non-ASCII (signed-char comparison)
- 8-bit C1 OSC/CSI (0x9D/0x9B) now initialize parser state identically
  to their ESC-sequence forms
- Skip PrCloseDoc after failed PrOpenDoc (printing builds)
- Finger: fix stray CR on chunk-split CR LF; cut UI-blocking fetch
  timeout from 10s to 3s (finger builds)
- prefs_save reports disk errors and flushes the volume instead of
  silently losing settings; Gestalt guarded in prefs_get_location
- Rename misnamed TCP_STATE_TIME_WAIT to TCP_STATE_CLOSE_WAIT

Performance:
- Window menu updates one item's text on OSC title changes instead of
  a full rebuild per output burst
- Session logging batches file writes (half-full/1s/stop/disconnect)
  instead of one synchronous FSWrite per TCP segment (logging builds)
- Skip post-keystroke echo-poll spin when multiple sessions are open
- Theme import saves prefs once, not twice (color builds)

All three presets build clean; migration test passes 112/112.

+++ src/color.c
@@ -138,7 +138,7 @@ color_nearest_256(unsigned char r, unsigned char g, unsigned char b)
 	short cr, cg, cb;
 	short gi;
 	short cube_idx;
-	short cube_dist, gray_dist;
+	long cube_dist, gray_dist;
 	short gray_val;
 	unsigned char gray_idx;
 
@@ -165,9 +165,9 @@ color_nearest_256(unsigned char r, unsigned char g, unsigned char b)
 	cube_idx = 16 + 36 * cr + 6 * cg + cb;
 
 	/* Distance to cube color (squared, per-component) */
-	cube_dist = (r - cube_vals[cr]) * (r - cube_vals[cr]) +
-	    (g - cube_vals[cg]) * (g - cube_vals[cg]) +
-	    (b - cube_vals[cb]) * (b - cube_vals[cb]);
+	cube_dist = (long)(r - cube_vals[cr]) * (r - cube_vals[cr]) +
+	    (long)(g - cube_vals[cg]) * (g - cube_vals[cg]) +
+	    (long)(b - cube_vals[cb]) * (b - cube_vals[cb]);
 
 	/* Check grayscale ramp (indices 232-255: 8, 18, 28, ..., 238) */
 	gray_val = (r + g + b) / 3;
@@ -182,9 +182,9 @@ color_nearest_256(unsigned char r, unsigned char g, unsigned char b)
 	/* Distance to grayscale value */
 	{
 		short gv = 8 + gi * 10;
-		gray_dist = (r - gv) * (r - gv) +
-		    (g - gv) * (g - gv) +
-		    (b - gv) * (b - gv);
+		gray_dist = (long)(r - gv) * (r - gv) +
+		    (long)(g - gv) * (g - gv) +
+		    (long)(b - gv) * (b - gv);
 	}
 
 	return (gray_dist < cube_dist) ? gray_idx :
